refactor: Move utils to a single folder
