+++ src/utils/cancelBubbleEventHandler.js
@@ -1,0 +1,5 @@
+export default handler => e => {
+  e.stopPropagation()
+  e.nativeEvent.stopImmediatePropagation()
+  handler(e)
+}
